修改完善music-client-vue功能（收藏...）、music-server修正接口

+++ music-server/src/main/java/club/jming/musicServer/service/CollectService.java
@@ -13,7 +13,7 @@ public interface CollectService extends IService<Collect> {
 
     boolean updateCollectMsg(Collect collect);
 
-    boolean deleteCollect(Integer userId, Integer MusicId);
+    boolean deleteCollect(Long userId, Long MusicId);
 
     List<Collect> allCollect();
